+++ test/community/training/easy/murderInTheVillage/CGTest.js
@@ -1,0 +1,107 @@
+import { assert } from 'chai';
+import sinon from 'sinon';
+import File from '../../../../File.js';
+import { execute } from '../../../../../lib/community/training/easy/murderInTheVillage/murderInTheVillage.js';
+
+const __dirname = new URL('.', import.meta.url).pathname;
+
+suite("Murder in the village!", function() {
+    const sandbox = sinon.createSandbox();
+
+    setup(function () {
+        sandbox.stub(console, "log");
+    });
+
+    teardown(function () {
+        sandbox.restore();
+    });
+
+
+    test("Test 1", function() {
+        let inputFile = new File(__dirname + 'input/01 - test 1.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "It was me!"
+        );
+    });
+
+    test("Test 2", function() {
+        let inputFile = new File(__dirname + 'input/02 - test 2.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Charles did it!"
+        );
+    });
+
+    test("Test 3", function() {
+        let inputFile = new File(__dirname + 'input/03 - test 3.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Charles did it!"
+        );
+    });
+
+    test("Test 4", function() {
+        let inputFile = new File(__dirname + 'input/04 - test 4.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Lance did it!"
+        );
+    });
+
+    test("Test 5", function() {
+        let inputFile = new File(__dirname + 'input/05 - test 5.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Ben did it!"
+        );
+    });
+
+    test("Test 6", function() {
+        let inputFile = new File(__dirname + 'input/06 - test 6.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Bob did it!"
+        );
+    });
+
+    test("Test 7", function() {
+        let inputFile = new File(__dirname + 'input/07 - test 7.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Hank did it!"
+        );
+    });
+
+    test("Test 8", function() {
+        let inputFile = new File(__dirname + 'input/08 - test 8.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            "Dylan did it!"
+        );
+    });
+});