+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Unique prefixes".
 - Tests for "The mystic rectangle".
 - Tests for "The travelling salesman problem".
+- Tests for "Murder in the village!".
 
 ## [1.7.0] - 2022-03-31
 ### Added